channels/timeless-clips/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "timeless-clips"
version = "0.1.0"
description = "Public domain clips reimagined as viral YouTube Shorts and TikTok content"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "AreteDriver"}]
keywords = ["youtube", "shorts", "tiktok", "public-domain", "internet-archive"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
tts = ["piper-tts>=1.0.0"]
captions = ["openai-whisper>=20230918"]

[project.scripts]
timeless-clips = "timeless_clips.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/timeless_clips/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["timeless_clips"]

[tool.coverage.report]
fail_under = 80
show_missing = true
